validator helpers isValid need revisions
validator helpers isValid need revisions. Rewrote for Ark and Valibot so it could work for now.
For Zod it's already ok because we validate instanceof ZodType.
For Joi, we validate joi properties so it's ok
TODO: ArkType and Valibot
We are missing good property check ot see if the schema is valid